философы научились понимать, когда им стоит прилечь, осталась проблема с тем, что потоки должны все отключаться, если хоть 1 умер

+++ philo/includes/philo_one.h
@@ -25,7 +25,7 @@ typedef struct	s_philo
 	int right_fork;
 	int state;
 	int times_eat;
-	int time_of_death;
+	unsigned long time_of_death;
 }	t_philo;
 
 typedef struct	s_all
